requirements.txt
anytree==2.8.0
graphviz==2.50.0
nltk==3.6.7
numpy==1.23.3
opencv-python==4.6.0.66
pandas==1.5.0
pillow==9.2.0
pip==22.2.2
requests==2.28.1
svglue==0.3.0
svgutils==0.3.4
tabulate==0.8.10
tk==8.6.12
python-statemachine==0.8.0